refactor: move TypeInfo subclasses to the module they are exported from
