+++ psycopg/psycopg/_typeinfo.py
@@ -7,7 +7,6 @@ information to the adapters if needed.
 
 # Copyright (C) 2020 The Psycopg Team
 
-from enum import Enum
 from typing import Any, Dict, Iterator, Optional, overload
 from typing import Sequence, Tuple, Type, TypeVar, Union, TYPE_CHECKING
 from typing_extensions import TypeAlias
@@ -188,181 +187,6 @@ ORDER BY t.oid
     def _added(self, registry: "TypesRegistry") -> None:
         """Method called by the `!registry` when the object is added there."""
         pass
-
-
-class RangeInfo(TypeInfo):
-    """Manage information about a range type."""
-
-    __module__ = "psycopg.types.range"
-
-    def __init__(
-        self,
-        name: str,
-        oid: int,
-        array_oid: int,
-        *,
-        regtype: str = "",
-        subtype_oid: int,
-    ):
-        super().__init__(name, oid, array_oid, regtype=regtype)
-        self.subtype_oid = subtype_oid
-
-    @classmethod
-    def _get_info_query(cls, conn: "BaseConnection[Any]") -> Query:
-        from .sql import SQL
-
-        return SQL(
-            """\
-SELECT t.typname AS name, t.oid AS oid, t.typarray AS array_oid,
-    t.oid::regtype::text AS regtype,
-    r.rngsubtype AS subtype_oid
-FROM pg_type t
-JOIN pg_range r ON t.oid = r.rngtypid
-WHERE t.oid = {regtype}
-"""
-        ).format(regtype=cls._to_regtype(conn))
-
-    def _added(self, registry: "TypesRegistry") -> None:
-        # Map ranges subtypes to info
-        registry._registry[RangeInfo, self.subtype_oid] = self
-
-
-class MultirangeInfo(TypeInfo):
-    """Manage information about a multirange type."""
-
-    __module__ = "psycopg.types.multirange"
-
-    def __init__(
-        self,
-        name: str,
-        oid: int,
-        array_oid: int,
-        *,
-        regtype: str = "",
-        range_oid: int,
-        subtype_oid: int,
-    ):
-        super().__init__(name, oid, array_oid, regtype=regtype)
-        self.range_oid = range_oid
-        self.subtype_oid = subtype_oid
-
-    @classmethod
-    def _get_info_query(cls, conn: "BaseConnection[Any]") -> Query:
-        from .sql import SQL
-
-        if conn.info.server_version < 140000:
-            raise e.NotSupportedError(
-                "multirange types are only available from PostgreSQL 14"
-            )
-
-        return SQL(
-            """\
-SELECT t.typname AS name, t.oid AS oid, t.typarray AS array_oid,
-    t.oid::regtype::text AS regtype,
-    r.rngtypid AS range_oid, r.rngsubtype AS subtype_oid
-FROM pg_type t
-JOIN pg_range r ON t.oid = r.rngmultitypid
-WHERE t.oid = {regtype}
-"""
-        ).format(regtype=cls._to_regtype(conn))
-
-    def _added(self, registry: "TypesRegistry") -> None:
-        # Map multiranges ranges and subtypes to info
-        registry._registry[MultirangeInfo, self.range_oid] = self
-        registry._registry[MultirangeInfo, self.subtype_oid] = self
-
-
-class CompositeInfo(TypeInfo):
-    """Manage information about a composite type."""
-
-    __module__ = "psycopg.types.composite"
-
-    def __init__(
-        self,
-        name: str,
-        oid: int,
-        array_oid: int,
-        *,
-        regtype: str = "",
-        field_names: Sequence[str],
-        field_types: Sequence[int],
-    ):
-        super().__init__(name, oid, array_oid, regtype=regtype)
-        self.field_names = field_names
-        self.field_types = field_types
-        # Will be set by register() if the `factory` is a type
-        self.python_type: Optional[type] = None
-
-    @classmethod
-    def _get_info_query(cls, conn: "BaseConnection[Any]") -> Query:
-        from .sql import SQL
-
-        return SQL(
-            """\
-SELECT
-    t.typname AS name, t.oid AS oid, t.typarray AS array_oid,
-    t.oid::regtype::text AS regtype,
-    coalesce(a.fnames, '{{}}') AS field_names,
-    coalesce(a.ftypes, '{{}}') AS field_types
-FROM pg_type t
-LEFT JOIN (
-    SELECT
-        attrelid,
-        array_agg(attname) AS fnames,
-        array_agg(atttypid) AS ftypes
-    FROM (
-        SELECT a.attrelid, a.attname, a.atttypid
-        FROM pg_attribute a
-        JOIN pg_type t ON t.typrelid = a.attrelid
-        WHERE t.oid = {regtype}
-        AND a.attnum > 0
-        AND NOT a.attisdropped
-        ORDER BY a.attnum
-    ) x
-    GROUP BY attrelid
-) a ON a.attrelid = t.typrelid
-WHERE t.oid = {regtype}
-"""
-        ).format(regtype=cls._to_regtype(conn))
-
-
-class EnumInfo(TypeInfo):
-    """Manage information about an enum type."""
-
-    __module__ = "psycopg.types.enum"
-
-    def __init__(
-        self,
-        name: str,
-        oid: int,
-        array_oid: int,
-        labels: Sequence[str],
-    ):
-        super().__init__(name, oid, array_oid)
-        self.labels = labels
-        # Will be set by register_enum()
-        self.enum: Optional[Type[Enum]] = None
-
-    @classmethod
-    def _get_info_query(cls, conn: "BaseConnection[Any]") -> Query:
-        from .sql import SQL
-
-        return SQL(
-            """\
-SELECT name, oid, array_oid, array_agg(label) AS labels
-FROM (
-    SELECT
-        t.typname AS name, t.oid AS oid, t.typarray AS array_oid,
-        e.enumlabel AS label
-    FROM pg_type t
-    LEFT JOIN  pg_enum e
-    ON e.enumtypid = t.oid
-    WHERE t.oid = {regtype}
-    ORDER BY e.enumsortorder
-) x
-GROUP BY name, oid, array_oid
-"""
-        ).format(regtype=cls._to_regtype(conn))
 
 
 class TypesRegistry:
